[更新] 修改TCIDAuthViewController中，出生日期不正确的bug [更新] TCHomeViewController中，所有网络加载后都要判断是否还有更多 [更新] TCWalletViewController中，点击对账单按钮时，将WalletID传过去

+++ individual/individual/Feature/Tools/Controller/TCRepairsDetailViewController.m
@@ -246,7 +246,7 @@ TCDatePickerViewDelegate>
 
 - (void)didClickConfirmButtonInDatePickerView:(TCDatePickerView *)view {
     NSTimeInterval timestamp = [view.datePicker.date timeIntervalSince1970];
-    self.repairsInfo.appointTime = (NSInteger)(timestamp * 1000);
+    self.repairsInfo.appointTime = timestamp * 1000;
     [self.tableView reloadData];
 }
